requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pydantic==2.9.2
python-multipart==0.0.12
pandas>=2.2.0
openpyxl>=3.1.5
python-dateutil>=2.9.0
fpdf2>=2.7.0
Pillow>=10.0.0
python-pptx>=1.0.2
requests>=2.31.0
bcrypt>=4.0.1
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
matplotlib>=3.5.0
numpy>=1.21.0
psutil>=5.9.0
schedule>=1.2.0
gunicorn>=21.2.0
